various fixes to get a working OS X build (rebb/TRSi)

+++ sound_emul/machines/CPU/z80.h
@@ -12,17 +12,17 @@ enum CpuType {NMOS, CMOS, BM1};
 class z80cpu {
 
 public:
-	int inputPorts[0x10000];
-	int inputPortsShort[0x100];
-	int outputPorts[0x10000];
-	int outputPortsShort[0x100];
-	
+	int8_t inputPorts[0x10000];
+	int8_t inputPortsShort[0x100];
+	int8_t outputPorts[0x10000];
+	int8_t outputPortsShort[0x100];
+
 	void reset();
-	
+
 	void setPC(int startAddress);
 	void setSP(int address);
 	int getPC();
-	
+
 	void execute_cycle();
 	void execute_debug();
 	int non_maskable_interrupt();
@@ -30,14 +30,14 @@ public:
 
 	z80cpu(int *mem, int z80Type);
 	~z80cpu();
-	
+
 private:
 	int *memory;
-	
+
 	int instructionCycles;
-	
+
 	int cpuType;
-	
+
 	int regA;
 	int regB;
 	int regC;
@@ -62,20 +62,20 @@ private:
 	int regIYL;
 	int regSP;
 	int regPC;
-	
+
 	int regQ;
 	int regMEMPTR;
 	bool regIFF1;
 	bool regIFF2;
 	bool interruptsEnabled;
 	int interruptMode;
-	
+
 	int regDummy;	//dummy register for generic instructions
-	
+
 	int *const regPtr[8] = {&regB, &regC, &regD, &regE, &regH, &regL, &regDummy, &regA};
-	
+
 	static const int conditionCodes[4];
-	
+
 	int parity_tbl[256];
 
 //generic 16-bit register instructions
@@ -101,7 +101,7 @@ private:
 	int xor_r8();
 	int or_r8();
 	int cp_r8();
-	
+
 //generic conditional instructions
 	int jr_cc_n();
 	int ret_cc();
@@ -110,23 +110,23 @@ private:
 
 //misc generic instructions
 	int rst_xx();
-	
+
 //regular (non-prefixed) instructions
-	
+
 	int nop();
 	int ld_atbc_a();
 	int rlca();
 	int ex_af_af();
 	int ld_a_atbc();
 	int rrca();
-	
+
 	int djnz();
 	int ld_atde_a();
 	int rla();
 	int jr();
 	int ld_a_atde();
 	int rra();
-	
+
 	int ld_atnn_hl();
 	int daa();
 	int ld_hl_atnn();
@@ -143,12 +143,12 @@ private:
 	int ld_a_atnn();
 	int dec_sp();
 	int ccf();
-	
+
 	int halt();
-	
+
 	int adx_a_athl();
 
-	int sbx_a_athl();	
+	int sbx_a_athl();
 
 	int and_athl();
 	int xor_athl();
@@ -161,20 +161,20 @@ private:
 	int prefix_cb();
 	int call_nn();
 	int adx_a_n();
-	
+
 	int out_atn_a();
 	int exx();
 	int in_a_atn();
 	int prefix_dd();
 	int sbx_a_n();
-	
+
 	int ex_atsp_hl();
 	int and_n();
 	int jp_athl();
 	int ex_de_hl();
 	int prefix_ed();
 	int xor_n();
-	
+
 	int pop_af();
 	int di();
 	int push_af();
@@ -200,20 +200,20 @@ private:
 	int ld_i_a();
 	int reti();
 	int ld_r_a();
-	
+
 	int im1();
 	int ld_a_i();
 	int im2();
 	int ld_a_r();
-	
+
 	int rrd();
 	int rld();
-	
+
 	int sbc_hl_sp();
 	int ld_atnn_sp();
 	int adc_hl_sp();
 	int ld_sp_atnn();
-	
+
 	int ldi();
 	int cpi();
 	int ini();
@@ -222,7 +222,7 @@ private:
 	int cpd();
 	int ind();
 	int outd();
-	
+
 	int ldir();
 	int cpir();
 	int inir();
@@ -231,7 +231,7 @@ private:
 	int cpdr();
 	int indr();
 	int otdr();
-	
+
 //prefix CB
 
 	int rlc_r8();
@@ -243,10 +243,10 @@ private:
 	int rl_athl();
 	int rr_r8();
 	int rr_athl();
-	
+
 	int sra_r8();
 	int sra_athl();
-	
+
 	int slx_r8();
 	int slx_athl();
 	int srl_r8();
@@ -254,10 +254,10 @@ private:
 
  	int bit_x_r8();
 	int bit_x_athl();
-	
+
  	int res_x_r8();
 	int res_x_athl();
-	
+
 	int set_x_r8();
 	int set_x_athl();
 
@@ -267,7 +267,7 @@ private:
 	int ld_r8_atixpd();
 	int ld_r8_ixh();
 	int ld_r8_ixl();
-	
+
 	int ld_ix_nn();
 	int ld_atnn_ix();
 	int inc_ix();
@@ -280,57 +280,57 @@ private:
 	int inc_ixl();
 	int dec_ixl();
 	int ld_ixl_n();
-	
+
 	int inc_atixpd();
 	int dec_atixpd();
 	int ld_atixpd_n();
 	int add_ix_sp();
-	
+
 	int ld_ixh_r8();
 	int ld_ixh_ixl();
 	int ld_ixl_r8();
 	int ld_ixl_ixh();
-	
+
 	int ld_atixpd_r8();
-	
+
 	int adx_a_ixh();
 	int adx_a_ixl();
 	int adx_a_atixpd();
-	
+
 	int sbx_a_ixh();
 	int sbx_a_ixl();
 	int sbx_a_atixpd();
-	
+
 	int and_ixh();
 	int and_ixl();
 	int and_atixpd();
 	int xor_ixh();
 	int xor_ixl();
 	int xor_atixpd();
-	
+
 	int or_ixh();
 	int or_ixl();
 	int or_atixpd();
 	int cp_ixh();
 	int cp_ixl();
 	int cp_atixpd();
-	
+
 	int prefix_ddcb();
-	
+
 	int pop_ix();
 	int ex_atsp_ix();
 	int push_ix();
 	int jp_atix();
-	
+
 	int ld_sp_ix();
-	
-	
+
+
 //prefix FD
 	int add_iy_r16();
 	int ld_r8_atiypd();
 	int ld_r8_iyh();
 	int ld_r8_iyl();
-	
+
 	int ld_iy_nn();
 	int ld_atnn_iy();
 	int inc_iy();
@@ -343,50 +343,50 @@ private:
 	int inc_iyl();
 	int dec_iyl();
 	int ld_iyl_n();
-	
+
 	int inc_atiypd();
 	int dec_atiypd();
 	int ld_atiypd_n();
 	int add_iy_sp();
-	
+
 	int ld_iyh_r8();
 	int ld_iyh_iyl();
 	int ld_iyl_r8();
 	int ld_iyl_iyh();
-	
+
 	int ld_atiypd_r8();
-	
+
 	int adx_a_iyh();
 	int adx_a_iyl();
 	int adx_a_atiypd();
-	
+
 	int sbx_a_iyh();
 	int sbx_a_iyl();
 	int sbx_a_atiypd();
-	
+
 	int and_iyh();
 	int and_iyl();
 	int and_atiypd();
 	int xor_iyh();
 	int xor_iyl();
 	int xor_atiypd();
-	
+
 	int or_iyh();
 	int or_iyl();
 	int or_atiypd();
 	int cp_iyh();
 	int cp_iyl();
 	int cp_atiypd();
-	
+
 	int prefix_fdcb();
-	
+
 	int pop_iy();
 	int ex_atsp_iy();
 	int push_iy();
 	int jp_atiy();
-	
+
 	int ld_sp_iy();
-	
+
 //prefix DDCB
 	int rlc_atixpd();
 	int rrc_atixpd();
@@ -396,11 +396,11 @@ private:
 	int sra_atixpd();
 	int sll_atixpd();
 	int srl_atixpd();
-	
+
 	int bit_x_atixpd();
 	int res_x_atixpd();
 	int set_x_atixpd();
-	
+
 //prefix FDCB
 	int rlc_atiypd();
 	int rrc_atiypd();
@@ -410,17 +410,17 @@ private:
 	int sra_atiypd();
 	int sll_atiypd();
 	int srl_atiypd();
-	
+
 	int bit_x_atiypd();
 	int res_x_atiypd();
-	int set_x_atiypd();		
-	
-	
+	int set_x_atiypd();
+
+
 	using CpuFun = decltype(mem_fn(&z80cpu::nop));
-	
+
 	//prefix FDCB, rotate/shift
  	const array<CpuFun, 8> cpu_instructions_fdcb {{
- 	
+
  		mem_fn(&z80cpu::rlc_atiypd),
  		mem_fn(&z80cpu::rrc_atiypd),
 
@@ -433,10 +433,10 @@ private:
 		mem_fn(&z80cpu::sll_atiypd),
 		mem_fn(&z80cpu::srl_atiypd)
  	}};
-	
+
 	//prefix DDCB, rotate/shift
  	const array<CpuFun, 8> cpu_instructions_ddcb {{
- 	
+
  		mem_fn(&z80cpu::rlc_atixpd),
  		mem_fn(&z80cpu::rrc_atixpd),
 
@@ -449,7 +449,7 @@ private:
 		mem_fn(&z80cpu::sll_atixpd),
 		mem_fn(&z80cpu::srl_atixpd)
  	}};
-	
+
 	//prefix FD
 	const array<CpuFun, 256> cpu_instructions_fd {{
 
@@ -469,7 +469,7 @@ private:
 		mem_fn(&z80cpu::dec_r8),
 		mem_fn(&z80cpu::ld_r8_n),
 		mem_fn(&z80cpu::rrca),
-		
+
 		mem_fn(&z80cpu::djnz),
 		mem_fn(&z80cpu::ld_r16_nn),
 		mem_fn(&z80cpu::ld_atde_a),
@@ -485,8 +485,8 @@ private:
 		mem_fn(&z80cpu::inc_r8),
 		mem_fn(&z80cpu::dec_r8),
 		mem_fn(&z80cpu::ld_r8_n),
-		mem_fn(&z80cpu::rra),		
-	
+		mem_fn(&z80cpu::rra),
+
 		mem_fn(&z80cpu::jr_cc_n),
 		mem_fn(&z80cpu::ld_iy_nn),
 		mem_fn(&z80cpu::ld_atnn_iy),
@@ -503,7 +503,7 @@ private:
 		mem_fn(&z80cpu::dec_iyl),
 		mem_fn(&z80cpu::ld_iyl_n),
 		mem_fn(&z80cpu::cpl),
-		
+
 		mem_fn(&z80cpu::jr_cc_n),
 		mem_fn(&z80cpu::ld_sp_nn),
 		mem_fn(&z80cpu::ld_atnn_a),
@@ -520,7 +520,7 @@ private:
 		mem_fn(&z80cpu::dec_r8),
 		mem_fn(&z80cpu::ld_r8_n),
 		mem_fn(&z80cpu::ccf),
-		
+
 		mem_fn(&z80cpu::nop),		//ld b,b
 		mem_fn(&z80cpu::ld_r8_r8),
 		mem_fn(&z80cpu::ld_r8_r8),
@@ -537,7 +537,7 @@ private:
 		mem_fn(&z80cpu::ld_r8_iyl),
 		mem_fn(&z80cpu::ld_r8_atiypd),
 		mem_fn(&z80cpu::ld_r8_r8),
-		
+
 		mem_fn(&z80cpu::ld_r8_r8),
 		mem_fn(&z80cpu::ld_r8_r8),
 		mem_fn(&z80cpu::nop),		//ld d,d
@@ -554,7 +554,7 @@ private:
 		mem_fn(&z80cpu::ld_r8_iyl),
 		mem_fn(&z80cpu::ld_r8_atiypd),
 		mem_fn(&z80cpu::ld_r8_r8),
-		
+
 		mem_fn(&z80cpu::ld_iyh_r8),
 		mem_fn(&z80cpu::ld_iyh_r8),
 		mem_fn(&z80cpu::ld_iyh_r8),
@@ -571,7 +571,7 @@ private:
 		mem_fn(&z80cpu::nop),		//mem_fn(&z80cpu::ld_iyl_iyl(),
 		mem_fn(&z80cpu::ld_r8_atiypd),
 		mem_fn(&z80cpu::ld_iyl_r8),
-		
+
 		mem_fn(&z80cpu::ld_atiypd_r8),
 		mem_fn(&z80cpu::ld_atiypd_r8),
 		mem_fn(&z80cpu::ld_atiypd_r8),
@@ -588,7 +588,7 @@ private:
 		mem_fn(&z80cpu::ld_r8_iyl),
 		mem_fn(&z80cpu::ld_r8_atiypd),
 		mem_fn(&z80cpu::nop),		//ld a,a
-		
+
 		mem_fn(&z80cpu::adx_a_r8),
 		mem_fn(&z80cpu::adx_a_r8),
 		mem_fn(&z80cpu::adx_a_r8),
@@ -605,7 +605,7 @@ private:
 		mem_fn(&z80cpu::adx_a_iyl),
 		mem_fn(&z80cpu::adx_a_atiypd),
 		mem_fn(&z80cpu::adx_a_r8),
-		
+
 		mem_fn(&z80cpu::sbx_a_r8),
 		mem_fn(&z80cpu::sbx_a_r8),
 		mem_fn(&z80cpu::sbx_a_r8),
@@ -622,7 +622,7 @@ private:
 		mem_fn(&z80cpu::sbx_a_iyl),
 		mem_fn(&z80cpu::sbx_a_atiypd),
 		mem_fn(&z80cpu::sbx_a_r8),
-		
+
 		mem_fn(&z80cpu::and_r8),
 		mem_fn(&z80cpu::and_r8),
 		mem_fn(&z80cpu::and_r8),
@@ -639,7 +639,7 @@ private:
 		mem_fn(&z80cpu::xor_iyl),
 		mem_fn(&z80cpu::xor_atiypd),
 		mem_fn(&z80cpu::xor_r8),
-		
+
 		mem_fn(&z80cpu::or_r8),
 		mem_fn(&z80cpu::or_r8),
 		mem_fn(&z80cpu::or_r8),
@@ -656,7 +656,7 @@ private:
 		mem_fn(&z80cpu::cp_iyl),
 		mem_fn(&z80cpu::cp_atiypd),
 		mem_fn(&z80cpu::cp_r8),
-		
+
 		mem_fn(&z80cpu::ret_cc),
 		mem_fn(&z80cpu::pop_r16),
 		mem_fn(&z80cpu::jp_cc_nn),
@@ -673,7 +673,7 @@ private:
 		mem_fn(&z80cpu::call_nn),
 		mem_fn(&z80cpu::adx_a_n),
 		mem_fn(&z80cpu::rst_xx),
-		
+
 		mem_fn(&z80cpu::ret_cc),
 		mem_fn(&z80cpu::pop_r16),
 		mem_fn(&z80cpu::jp_cc_nn),
@@ -690,7 +690,7 @@ private:
 		mem_fn(&z80cpu::prefix_dd),
 		mem_fn(&z80cpu::sbx_a_n),
 		mem_fn(&z80cpu::rst_xx),
-		
+
 		mem_fn(&z80cpu::ret_cc),
 		mem_fn(&z80cpu::pop_iy),
 		mem_fn(&z80cpu::jp_cc_nn),
@@ -707,7 +707,7 @@ private:
 		mem_fn(&z80cpu::prefix_ed),
 		mem_fn(&z80cpu::xor_n),
 		mem_fn(&z80cpu::rst_xx),
-		
+
 		mem_fn(&z80cpu::ret_cc),
 		mem_fn(&z80cpu::pop_af),
 		mem_fn(&z80cpu::jp_cc_nn),
@@ -723,9 +723,9 @@ private:
 		mem_fn(&z80cpu::call_cc_nn),
 		mem_fn(&z80cpu::prefix_fd),
 		mem_fn(&z80cpu::cp_n),
-		mem_fn(&z80cpu::rst_xx)		
+		mem_fn(&z80cpu::rst_xx)
 	}};
-	
+
 	//prefix DD
 	const array<CpuFun, 256> cpu_instructions_dd {{
 
@@ -745,7 +745,7 @@ private:
 		mem_fn(&z80cpu::dec_r8),
 		mem_fn(&z80cpu::ld_r8_n),
 		mem_fn(&z80cpu::rrca),
-		
+
 		mem_fn(&z80cpu::djnz),
 		mem_fn(&z80cpu::ld_r16_nn),
 		mem_fn(&z80cpu::ld_atde_a),
@@ -761,8 +761,8 @@ private:
 		mem_fn(&z80cpu::inc_r8),
 		mem_fn(&z80cpu::dec_r8),
 		mem_fn(&z80cpu::ld_r8_n),
-		mem_fn(&z80cpu::rra),		
-	
+		mem_fn(&z80cpu::rra),
+
 		mem_fn(&z80cpu::jr_cc_n),
 		mem_fn(&z80cpu::ld_ix_nn),
 		mem_fn(&z80cpu::ld_atnn_ix),
@@ -779,7 +779,7 @@ private:
 		mem_fn(&z80cpu::dec_ixl),
 		mem_fn(&z80cpu::ld_ixl_n),
 		mem_fn(&z80cpu::cpl),
-		
+
 		mem_fn(&z80cpu::jr_cc_n),
 		mem_fn(&z80cpu::ld_sp_nn),
 		mem_fn(&z80cpu::ld_atnn_a),
@@ -796,7 +796,7 @@ private:
 		mem_fn(&z80cpu::dec_r8),
 		mem_fn(&z80cpu::ld_r8_n),
 		mem_fn(&z80cpu::ccf),
-		
+
 		mem_fn(&z80cpu::nop),		//ld b,b
 		mem_fn(&z80cpu::ld_r8_r8),
 		mem_fn(&z80cpu::ld_r8_r8),
@@ -813,7 +813,7 @@ private:
 		mem_fn(&z80cpu::ld_r8_ixl),
 		mem_fn(&z80cpu::ld_r8_atixpd),
 		mem_fn(&z80cpu::ld_r8_r8),
-		
+
 		mem_fn(&z80cpu::ld_r8_r8),
 		mem_fn(&z80cpu::ld_r8_r8),
 		mem_fn(&z80cpu::nop),		//ld d,d
@@ -830,7 +830,7 @@ private:
 		mem_fn(&z80cpu::ld_r8_ixl),
 		mem_fn(&z80cpu::ld_r8_atixpd),
 		mem_fn(&z80cpu::ld_r8_r8),
-		
+
 		mem_fn(&z80cpu::ld_ixh_r8),
 		mem_fn(&z80cpu::ld_ixh_r8),
 		mem_fn(&z80cpu::ld_ixh_r8),
@@ -847,7 +847,7 @@ private:
 		mem_fn(&z80cpu::nop),		//mem_fn(&z80cpu::ld_ixl_ixl(),
 		mem_fn(&z80cpu::ld_r8_atixpd),
 		mem_fn(&z80cpu::ld_ixl_r8),
-		
+
 		mem_fn(&z80cpu::ld_atixpd_r8),
 		mem_fn(&z80cpu::ld_atixpd_r8),
 		mem_fn(&z80cpu::ld_atixpd_r8),
@@ -864,7 +864,7 @@ private:
 		mem_fn(&z80cpu::ld_r8_ixl),
 		mem_fn(&z80cpu::ld_r8_atixpd),
 		mem_fn(&z80cpu::nop),		//ld a,a
-		
+
 		mem_fn(&z80cpu::adx_a_r8),
 		mem_fn(&z80cpu::adx_a_r8),
 		mem_fn(&z80cpu::adx_a_r8),
@@ -881,7 +881,7 @@ private:
 		mem_fn(&z80cpu::adx_a_ixl),
 		mem_fn(&z80cpu::adx_a_atixpd),
 		mem_fn(&z80cpu::adx_a_r8),
-		
+
 		mem_fn(&z80cpu::sbx_a_r8),
 		mem_fn(&z80cpu::sbx_a_r8),
 		mem_fn(&z80cpu::sbx_a_r8),
@@ -898,7 +898,7 @@ private:
 		mem_fn(&z80cpu::sbx_a_ixl),
 		mem_fn(&z80cpu::sbx_a_atixpd),
 		mem_fn(&z80cpu::sbx_a_r8),
-		
+
 		mem_fn(&z80cpu::and_r8),
 		mem_fn(&z80cpu::and_r8),
 		mem_fn(&z80cpu::and_r8),
@@ -915,7 +915,7 @@ private:
 		mem_fn(&z80cpu::xor_ixl),
 		mem_fn(&z80cpu::xor_atixpd),
 		mem_fn(&z80cpu::xor_r8),
-		
+
 		mem_fn(&z80cpu::or_r8),
 		mem_fn(&z80cpu::or_r8),
 		mem_fn(&z80cpu::or_r8),
@@ -932,7 +932,7 @@ private:
 		mem_fn(&z80cpu::cp_ixl),
 		mem_fn(&z80cpu::cp_atixpd),
 		mem_fn(&z80cpu::cp_r8),
-		
+
 		mem_fn(&z80cpu::ret_cc),
 		mem_fn(&z80cpu::pop_r16),
 		mem_fn(&z80cpu::jp_cc_nn),
@@ -949,7 +949,7 @@ private:
 		mem_fn(&z80cpu::call_nn),
 		mem_fn(&z80cpu::adx_a_n),
 		mem_fn(&z80cpu::rst_xx),
-		
+
 		mem_fn(&z80cpu::ret_cc),
 		mem_fn(&z80cpu::pop_r16),
 		mem_fn(&z80cpu::jp_cc_nn),
@@ -966,7 +966,7 @@ private:
 		mem_fn(&z80cpu::prefix_dd),
 		mem_fn(&z80cpu::sbx_a_n),
 		mem_fn(&z80cpu::rst_xx),
-		
+
 		mem_fn(&z80cpu::ret_cc),
 		mem_fn(&z80cpu::pop_ix),
 		mem_fn(&z80cpu::jp_cc_nn),
@@ -983,7 +983,7 @@ private:
 		mem_fn(&z80cpu::prefix_ed),
 		mem_fn(&z80cpu::xor_n),
 		mem_fn(&z80cpu::rst_xx),
-		
+
 		mem_fn(&z80cpu::ret_cc),
 		mem_fn(&z80cpu::pop_af),
 		mem_fn(&z80cpu::jp_cc_nn),
@@ -999,12 +999,12 @@ private:
 		mem_fn(&z80cpu::call_cc_nn),
 		mem_fn(&z80cpu::prefix_fd),
 		mem_fn(&z80cpu::cp_n),
-		mem_fn(&z80cpu::rst_xx)		
+		mem_fn(&z80cpu::rst_xx)
 	}};
-	
+
 	//prefix CB, rotate/shift
 	const array<CpuFun, 64> cpu_instructions_cb {{
-	
+
 		mem_fn(&z80cpu::rlc_r8),
 		mem_fn(&z80cpu::rlc_r8),
 		mem_fn(&z80cpu::rlc_r8),
@@ -1038,7 +1038,7 @@ private:
 		mem_fn(&z80cpu::rr_r8),
 		mem_fn(&z80cpu::rr_athl),
 		mem_fn(&z80cpu::rr_r8),
-	
+
 		mem_fn(&z80cpu::slx_r8),
 		mem_fn(&z80cpu::slx_r8),
 		mem_fn(&z80cpu::slx_r8),
@@ -1055,7 +1055,7 @@ private:
 		mem_fn(&z80cpu::sra_r8),
 		mem_fn(&z80cpu::sra_athl),
 		mem_fn(&z80cpu::sra_r8),
-	
+
 		mem_fn(&z80cpu::slx_r8),
 		mem_fn(&z80cpu::slx_r8),
 		mem_fn(&z80cpu::slx_r8),
@@ -1076,7 +1076,7 @@ private:
 
 	//prefix CB, bit ops
 	const array<CpuFun, 8> cpu_instructions_cb_bit {{
-	
+
 		mem_fn(&z80cpu::bit_x_r8),
 		mem_fn(&z80cpu::bit_x_r8),
 		mem_fn(&z80cpu::bit_x_r8),
@@ -1086,10 +1086,10 @@ private:
 		mem_fn(&z80cpu::bit_x_athl),
 		mem_fn(&z80cpu::bit_x_r8)
 	}};
-	
+
 	//prefix CB, res ops
 	const array<CpuFun, 8> cpu_instructions_cb_res {{
-	
+
 		mem_fn(&z80cpu::res_x_r8),
 		mem_fn(&z80cpu::res_x_r8),
 		mem_fn(&z80cpu::res_x_r8),
@@ -1099,10 +1099,10 @@ private:
 		mem_fn(&z80cpu::res_x_athl),
 		mem_fn(&z80cpu::res_x_r8)
 	}};
-	
+
 	//prefix CB, set ops
 	const array<CpuFun, 8> cpu_instructions_cb_set {{
-	
+
 		mem_fn(&z80cpu::set_x_r8),
 		mem_fn(&z80cpu::set_x_r8),
 		mem_fn(&z80cpu::set_x_r8),
@@ -1112,7 +1112,7 @@ private:
 		mem_fn(&z80cpu::set_x_athl),
 		mem_fn(&z80cpu::set_x_r8)
 	}};
-	
+
 	//prefix ED
 	const array<CpuFun, 124> cpu_instructions_ed {{
 
@@ -1132,7 +1132,7 @@ private:
 		mem_fn(&z80cpu::reti),
 		mem_fn(&z80cpu::im0),
 		mem_fn(&z80cpu::ld_r_a),
-	
+
 		mem_fn(&z80cpu::in_r8_atc),
 		mem_fn(&z80cpu::out_atc_r8),
 		mem_fn(&z80cpu::sbc_hl_r16),
@@ -1149,7 +1149,7 @@ private:
 		mem_fn(&z80cpu::reti),
 		mem_fn(&z80cpu::im2),
 		mem_fn(&z80cpu::ld_a_r),
-	
+
 		mem_fn(&z80cpu::in_r8_atc),
 		mem_fn(&z80cpu::out_atc_r8),
 		mem_fn(&z80cpu::sbc_hl_r16),
@@ -1166,7 +1166,7 @@ private:
 		mem_fn(&z80cpu::reti),
 		mem_fn(&z80cpu::im0),
 		mem_fn(&z80cpu::rld),
-	
+
 		mem_fn(&z80cpu::in_r8_atc),
 		mem_fn(&z80cpu::out_atc_r8),
 		mem_fn(&z80cpu::sbc_hl_sp),
@@ -1183,7 +1183,7 @@ private:
 		mem_fn(&z80cpu::reti),
 		mem_fn(&z80cpu::im2),
 		mem_fn(&z80cpu::nop),
-	
+
 		mem_fn(&z80cpu::nop),
 		mem_fn(&z80cpu::nop),
 		mem_fn(&z80cpu::nop),
@@ -1200,7 +1200,7 @@ private:
 		mem_fn(&z80cpu::nop),
 		mem_fn(&z80cpu::nop),
 		mem_fn(&z80cpu::nop),
-		
+
 		mem_fn(&z80cpu::nop),
 		mem_fn(&z80cpu::nop),
 		mem_fn(&z80cpu::nop),
@@ -1217,7 +1217,7 @@ private:
 		mem_fn(&z80cpu::nop),
 		mem_fn(&z80cpu::nop),
 		mem_fn(&z80cpu::nop),
-	
+
 		mem_fn(&z80cpu::ldi),
 		mem_fn(&z80cpu::cpi),
 		mem_fn(&z80cpu::ini),
@@ -1234,7 +1234,7 @@ private:
 		mem_fn(&z80cpu::nop),
 		mem_fn(&z80cpu::nop),
 		mem_fn(&z80cpu::nop),
-	
+
 		mem_fn(&z80cpu::ldir),
 		mem_fn(&z80cpu::cpir),
 		mem_fn(&z80cpu::inir),
@@ -1249,8 +1249,8 @@ private:
 		mem_fn(&z80cpu::otdr)
 	}};
 
-	
-//non-prefixed	
+
+//non-prefixed
 	const array<CpuFun, 256> cpu_instructions {{
 		mem_fn(&z80cpu::nop),
 	 	mem_fn(&z80cpu::ld_r16_nn),
@@ -1268,7 +1268,7 @@ private:
 		mem_fn(&z80cpu::dec_r8),
 		mem_fn(&z80cpu::ld_r8_n),
 		mem_fn(&z80cpu::rrca),
-		
+
 		mem_fn(&z80cpu::djnz),
 		mem_fn(&z80cpu::ld_r16_nn),
 		mem_fn(&z80cpu::ld_atde_a),
@@ -1285,7 +1285,7 @@ private:
 		mem_fn(&z80cpu::dec_r8),
 		mem_fn(&z80cpu::ld_r8_n),
 		mem_fn(&z80cpu::rra),
-		
+
 		mem_fn(&z80cpu::jr_cc_n),
 		mem_fn(&z80cpu::ld_r16_nn),
 		mem_fn(&z80cpu::ld_atnn_hl),
@@ -1302,7 +1302,7 @@ private:
 		mem_fn(&z80cpu::dec_r8),
 		mem_fn(&z80cpu::ld_r8_n),
 		mem_fn(&z80cpu::cpl),
-		
+
 		mem_fn(&z80cpu::jr_cc_n),
 		mem_fn(&z80cpu::ld_sp_nn),
 		mem_fn(&z80cpu::ld_atnn_a),
@@ -1319,7 +1319,7 @@ private:
 		mem_fn(&z80cpu::dec_r8),
 		mem_fn(&z80cpu::ld_r8_n),
 		mem_fn(&z80cpu::ccf),
-		
+
 		mem_fn(&z80cpu::nop),		//ld b,b
 		mem_fn(&z80cpu::ld_r8_r8),
 		mem_fn(&z80cpu::ld_r8_r8),
@@ -1336,7 +1336,7 @@ private:
 		mem_fn(&z80cpu::ld_r8_r8),
 		mem_fn(&z80cpu::ld_r8_athl),
 		mem_fn(&z80cpu::ld_r8_r8),
-		
+
 		mem_fn(&z80cpu::ld_r8_r8),
 		mem_fn(&z80cpu::ld_r8_r8),
 		mem_fn(&z80cpu::nop),		//ld d,d
@@ -1353,7 +1353,7 @@ private:
 		mem_fn(&z80cpu::ld_r8_r8),
 		mem_fn(&z80cpu::ld_r8_athl),
 		mem_fn(&z80cpu::ld_r8_r8),
-		
+
 		mem_fn(&z80cpu::ld_r8_r8),
 		mem_fn(&z80cpu::ld_r8_r8),
 		mem_fn(&z80cpu::ld_r8_r8),
@@ -1370,7 +1370,7 @@ private:
 		mem_fn(&z80cpu::nop),		//ld l,l
 		mem_fn(&z80cpu::ld_r8_athl),
 		mem_fn(&z80cpu::ld_r8_r8),
-		
+
 		mem_fn(&z80cpu::ld_athl_r8),
 		mem_fn(&z80cpu::ld_athl_r8),
 		mem_fn(&z80cpu::ld_athl_r8),
@@ -1387,7 +1387,7 @@ private:
 		mem_fn(&z80cpu::ld_r8_r8),
 		mem_fn(&z80cpu::ld_r8_athl),
 		mem_fn(&z80cpu::nop),		//ld a,a
-		
+
 		mem_fn(&z80cpu::adx_a_r8),
 		mem_fn(&z80cpu::adx_a_r8),
 		mem_fn(&z80cpu::adx_a_r8),
@@ -1404,7 +1404,7 @@ private:
 		mem_fn(&z80cpu::adx_a_r8),
 		mem_fn(&z80cpu::adx_a_athl),
 		mem_fn(&z80cpu::adx_a_r8),
-		
+
 		mem_fn(&z80cpu::sbx_a_r8),
 		mem_fn(&z80cpu::sbx_a_r8),
 		mem_fn(&z80cpu::sbx_a_r8),
@@ -1421,7 +1421,7 @@ private:
 		mem_fn(&z80cpu::sbx_a_r8),
 		mem_fn(&z80cpu::sbx_a_athl),
 		mem_fn(&z80cpu::sbx_a_r8),
-		
+
 		mem_fn(&z80cpu::and_r8),
 		mem_fn(&z80cpu::and_r8),
 		mem_fn(&z80cpu::and_r8),
@@ -1455,7 +1455,7 @@ private:
 		mem_fn(&z80cpu::cp_r8),
 		mem_fn(&z80cpu::cp_athl),
 		mem_fn(&z80cpu::cp_r8),
-		
+
 		mem_fn(&z80cpu::ret_cc),
 		mem_fn(&z80cpu::pop_r16),
 		mem_fn(&z80cpu::jp_cc_nn),
@@ -1472,7 +1472,7 @@ private:
 		mem_fn(&z80cpu::call_nn),
 		mem_fn(&z80cpu::adx_a_n),
 		mem_fn(&z80cpu::rst_xx),
-				
+
 		mem_fn(&z80cpu::ret_cc),
 		mem_fn(&z80cpu::pop_r16),
 		mem_fn(&z80cpu::jp_cc_nn),
@@ -1489,7 +1489,7 @@ private:
 		mem_fn(&z80cpu::prefix_dd),
 		mem_fn(&z80cpu::sbx_a_n),
 		mem_fn(&z80cpu::rst_xx),
-		
+
 		mem_fn(&z80cpu::ret_cc),
 		mem_fn(&z80cpu::pop_r16),
 		mem_fn(&z80cpu::jp_cc_nn),
@@ -1506,7 +1506,7 @@ private:
 		mem_fn(&z80cpu::prefix_ed),
 		mem_fn(&z80cpu::xor_n),
 		mem_fn(&z80cpu::rst_xx),
-		
+
 		mem_fn(&z80cpu::ret_cc),
 		mem_fn(&z80cpu::pop_af),
 		mem_fn(&z80cpu::jp_cc_nn),
